services/catalog-api/pyproject.toml
[project]
name = "catalog-api"
version = "0.1.0"
description = "Catalog API"
authors = [
    {name = "José Alberto Ayra",email = "jalbertoayra@outlook.es"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "alembic (>=1.18.0,<2.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "ruff (>=0.14.10,<0.15.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]

[tool.poetry]
name = "catalog-api"
version = "0.1.0"
description = "Catalog API"
authors = []
packages = [{ include = "app" }]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
